add support for LSF HPC cluster scheduling system

+++ scripts/pipeline/pantagruel_pipeline_06.1_HPC_full_ML_gene_trees.sh
@@ -31,9 +31,15 @@ source ${envsourcescript}
 mkdir -p ${ptglogs}/raxml/gene_trees/ ${$mlgenetrees}/
 
 ## compute first pass of gene trees with RAxML, using rapid bootstrap to estimate branch supports
-rm -f ${genetrees}/cdsfams_*_aln_list
-sortminsizes=($(for famset in `ls ${genetrees}/*cdsfams_*` ; do echo $famset | sed -e 's/.*cdsfams_minsize\([0-9]\+.*\)/\1/' ; done | sort -n))
-allcdsfam2phylo=($(for famminsize in ${sortminsizes[@]} ; do ls ${genetrees}/*${famminsize} ; done))
+if [[ "${resumetask}" == "true" && ! -z "$(ls ${genetrees}/cdsfams_*_aln_list 2> /dev/null)" ]] ; then
+  echo "will re-use previously established lists of gene families for which to compute trees" 
+  allcdsfam2phylo=($(ls ${genetrees}/cdsfams_*_aln_list))
+else
+  rm -f ${genetrees}/cdsfams_*_aln_list
+  sortminsizes=($(for famset in `ls ${genetrees}/*cdsfams_*` ; do echo $famset | sed -e 's/.*cdsfams_minsize\([0-9]\+.*\)/\1/' ; done | sort -n))
+  allcdsfam2phylo=($(for famminsize in ${sortminsizes[@]} ; do ls ${genetrees}/*${famminsize} ; done))
+fi
+  
 allmems=(4 8 32 64)
 allwalltimes=(12 24 48 72)
 allncpus=(4 4 4 16)
